requirements.txt
ad3 @ file:///D:/ad3-2.2.1-cp38-cp38-win_amd64.whl
aliyun-python-sdk-core==2.13.26
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.12.0
allure-pytest==2.8.24
allure-python-commons==2.8.24
appdirs==1.4.4
Appium-Python-Client==1.0.2
asgiref==3.2.10
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
bottle==0.12.18
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
config==0.5.0.post0
ConfigArgParse==1.2.3
constantly==15.1.0
crcmod==1.7
cryptography==3.0
cssselect==1.1.0
cycler==0.10.0
ddt==1.4.1
distlib==0.3.0
Django==3.1.1
document==1.0
empy==3.3.4
enum34==1.1.10
execute==1.2
fake-useragent==0.1.11
filelock==3.0.12
Flask==1.1.2
Flask-BasicAuth==0.2.0
future==0.18.2
gevent==20.6.2
geventhttpclient==1.4.4
greenlet==0.4.16
html-testRunner==1.2.1
hyperlink==20.0.1
idna==2.9
incremental==17.5.0
iniconfig==1.1.1
isort==4.3.21
itchat==1.3.10
itemadapter==0.1.0
itemloaders==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lml==0.0.9
locust==1.1.1
lxml==4.5.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
msgpack==1.0.0
mysqlclient==2.0.1
namedlist==1.8
ntplib==0.3.4
numpy==1.19.0
opencv-python==4.4.0.46
oss2==2.13.0
packaging==20.4
pandas==1.0.5
parameterized==0.7.4
parsel==1.6.0
Pillow==7.1.2
pipenv==2020.6.2
pluggy==0.13.1
Protego==0.1.16
psutil==5.7.2
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
PyDispatcher==2.0.5
pyexcel==0.6.4
pyexcel-io==0.5.20
pygame==1.9.6
PyHamcrest==2.0.2
pylint==2.5.3
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyStrich==0.8
pytesseract==0.3.4
pytest==6.1.2
python-dateutil==2.8.1
python-docx @ file:///D:/python_docx-0.8.10-py2.py3-none-any.whl
pytils==0.3
pytz==2020.1
pywin32==300
PyYAML==5.3.1
pyzmq==19.0.2
queuelib==1.5.0
requests==2.23.0
scipy==1.5.2
seaborn==0.10.1
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.1
sqlparse==0.3.1
texttable==1.6.2
toml==0.10.1
tools==0.1.9
torch @ file:///D:/torch-1.6.0-cp38-cp38-win_amd64.whl
torchvision @ file:///C:/Users/GUO/Desktop/torchvision-0.6.1%2Bcu92-cp38-cp38-win_amd64.whl
tqdm==4.46.1
unicodecsv==0.14.1
urllib3==1.25.9
virtualenv==20.0.24
virtualenv-clone==0.5.4
w3lib==1.22.0
Werkzeug==1.0.1
wrapt==1.12.1
wxPython==4.1.0
xlrd==1.2.0
xlwt==1.3.0
zope.event==4.4
zope.interface==5.1.0
